move HandshakePayload under clients/
