+++ packages/backend/src/api/factory.ts
@@ -7,7 +7,6 @@ import {
   ClientAPI,
   DomainAPI,
   EmailAddressAPI,
-  HandshakePayloadAPI,
   InstanceAPI,
   InvitationAPI,
   JwksAPI,
@@ -46,7 +45,6 @@ export function createBackendApiClient(options: CreateBackendApiOptions) {
     clients: new ClientAPI(request),
     domains: new DomainAPI(request),
     emailAddresses: new EmailAddressAPI(request),
-    handshakePayloads: new HandshakePayloadAPI(request),
     instance: new InstanceAPI(request),
     invitations: new InvitationAPI(request),
     jwks: new JwksAPI(request),
