+++ packages/backend/src/tokens/handshake.ts
@@ -1,4 +1,4 @@
-import type { HandshakePayloadAPI } from '../api/endpoints';
+import type { ClientAPI } from '../api/endpoints/ClientApi';
 import { constants, SUPPORTED_BAPI_VERSION } from '../constants';
 import { TokenVerificationError, TokenVerificationErrorAction, TokenVerificationErrorReason } from '../errors';
 import type { VerifyJwtOptions } from '../jwt';
@@ -88,18 +88,18 @@ export class HandshakeService {
   private readonly authenticateContext: AuthenticateContext;
   private readonly organizationMatcher: OrganizationMatcher;
   private readonly options: { organizationSyncOptions?: OrganizationSyncOptions };
-  private readonly handshakePayloadApi?: HandshakePayloadAPI;
+  private readonly clientApi?: ClientAPI;
 
   constructor(
     authenticateContext: AuthenticateContext,
     options: { organizationSyncOptions?: OrganizationSyncOptions },
     organizationMatcher: OrganizationMatcher,
-    handshakePayloadApi?: HandshakePayloadAPI,
+    clientApi?: ClientAPI,
   ) {
     this.authenticateContext = authenticateContext;
     this.options = options;
     this.organizationMatcher = organizationMatcher;
-    this.handshakePayloadApi = handshakePayloadApi;
+    this.clientApi = clientApi;
     this.redirectLoopCounter = 0;
   }
 
@@ -179,16 +179,18 @@ export class HandshakeService {
     const cookiesToSet: string[] = [];
 
     if (this.authenticateContext.handshakeNonce) {
-      if (!this.handshakePayloadApi) {
-        console.error('Clerk: HandshakeService: handshakePayloadApi is not available, cannot process handshake nonce.');
+      if (!this.clientApi) {
+        console.error('Clerk: HandshakeService: clientApi is not available, cannot process handshake nonce.');
       } else {
-        const handshakePayload = await this.handshakePayloadApi.getHandshakePayload({
+        const handshakePayload = await this.clientApi.getHandshakePayload({
           nonce: this.authenticateContext.handshakeNonce,
         });
         if (handshakePayload) {
-          console.log('Clerk: Handshake payload by nonce:', handshakePayload.payload);
+          console.log('Clerk: Handshake payload by nonce:', handshakePayload);
+          cookiesToSet.push(...handshakePayload.directives);
         }
       }
+      cookiesToSet.push(`${constants.Cookies.HandshakeNonce}=; Max-Age=0; Path=/; HttpOnly; SameSite=Lax`);
     }
     if (this.authenticateContext.handshakeToken) {
       const handshakePayload = await verifyHandshakeToken(
