+++ packages/backend/src/tokens/request.ts
@@ -89,7 +89,7 @@ export async function authenticateRequest(
     authenticateContext,
     { organizationSyncOptions: options.organizationSyncOptions },
     organizationMatcher,
-    options.apiClient?.handshakePayloads,
+    options.apiClient?.clients,
   );
 
   async function refreshToken(
